+++ src/core/engines/Cesium/Feature/utils.tsx
@@ -15,12 +15,14 @@ import {
   ImageryLayer,
 } from "cesium";
 import md5 from "js-md5";
+import { pick } from "lodash-es";
 import {
   ComponentProps,
   ComponentType,
   ForwardedRef,
   forwardRef,
   useLayoutEffect,
+  useMemo,
   useRef,
 } from "react";
 import { type CesiumComponentRef, Entity } from "resium";
@@ -226,4 +228,13 @@ export const generateIDWithMD5 = (id: string) => {
   hash.update(id);
 
   return hash.hex();
+};
+
+export const usePick = <T extends object, U extends keyof T>(
+  o: T | undefined | null,
+  fields: readonly U[],
+): Pick<T, U> | undefined => {
+  const p = useMemo(() => (o ? pick(o, fields) : undefined), [o, fields]);
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  return useMemo(() => p, [JSON.stringify(p)]);
 };
